adding skill object calculations and removing skill property from profile

+++ matchup/models.py
@@ -133,8 +133,8 @@ class Skill(models.Model):
     """
     player = models.ForeignKey(User)
     sport = models.CharField(max_length=40)
-    skill = models.FloatField(null=True, blank=True)
-    sportsmanship = models.FloatField(null=True, blank=True)
+    skill = models.DecimalField(null=True, blank=True, max_digits=5, decimal_places=2)
+    sportsmanship = models.DecimalField(null=True, blank=True, max_digits=5, decimal_places=2)
     num_feedbacks = models.IntegerField(null=True)
     created_at = models.DateTimeField(auto_now_add=True, null=True)
     updated_at = models.DateField(auto_now=True, null=True)
